feat(backend): agregar auth para validación de jwt, loggin, logout,
chore(backend): agregar protección a rutas por roles y validación de token que viene en las cookies, agregar configuración de cookies

+++ backend/src/note/note.controller.ts
@@ -6,35 +6,71 @@ import {
   Param,
   Patch,
   Post,
+  Req,
+  UnauthorizedException,
   UseFilters,
+  UseGuards,
 } from '@nestjs/common';
 import { NoteService } from './note.service';
-import { CreateNoteDto } from '././dtos/create-note.dto';
-import { UpdateNoteDto } from '././dtos/update-note.dto';
+import { CreateNoteDto } from './dtos/create-note.dto';
+import { UpdateNoteDto } from './dtos/update-note.dto';
 import { GlobalExceptionFilter } from '../common/filters/global-exception.filter';
+import { JwtAuthGuard } from '../auth/jwt-auth.guard';
+import { Request } from 'express';
 
+@UseGuards(JwtAuthGuard)
 @UseFilters(GlobalExceptionFilter)
 @Controller('note')
 export class NoteController {
   constructor(private readonly noteService: NoteService) {}
 
   @Post()
-  create(@Body() createNoteDto: CreateNoteDto) {
-    return this.noteService.create(createNoteDto);
+  create(@Body() createNoteDto: CreateNoteDto, @Req() req: Request) {
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-expect-error
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+    return this.noteService.create(createNoteDto, req.user.sub);
   }
 
   @Get(':directoryId')
-  findAll(@Param('directoryId') directoryId: string) {
-    return this.noteService.findAll(directoryId);
+  findAll(@Param('directoryId') directoryId: string, @Req() req: Request) {
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-expect-error
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+    return this.noteService.findAll(directoryId, req.user.sub);
   }
 
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateNoteDto: UpdateNoteDto) {
+  @Patch(':id/:userId')
+  update(
+    @Param('id') id: string,
+    @Param('userId') userId: string,
+    @Body() updateNoteDto: UpdateNoteDto,
+    @Req() req: Request,
+  ) {
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-expect-error
+    if (req.user.sub !== userId) {
+      throw new UnauthorizedException('Not authorized');
+    }
     return this.noteService.update(id, updateNoteDto);
   }
 
-  @Delete(':id')
-  remove(@Param('id') id: string) {
+  @Delete(':id/:userId')
+  remove(
+    @Param('id') id: string,
+    @Param('userId') userId: string,
+    @Req() req: Request,
+  ) {
+    console.log(userId);
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-expect-error
+    if (req.user.sub !== userId) {
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-expect-error
+      console.log(req.user.sub);
+      console.log(userId);
+      throw new UnauthorizedException('Not authorized');
+    }
     return this.noteService.remove(id);
   }
 }
